backend/requirements.txt
fastapi
uvicorn
pydantic-ai
pydantic-ai[google]
pypdf
python-multipart
python-dotenv
openai
